refactor: Merging into SkipBoardNoCheck
